Fix weather script on macos
In order to enable use of the weather script on MacOS, this commit makes
a few slight changes for an easier UX. The first is allowing everything
after the first two arguments to be treated as the third argument. This
is useful when we want to pass in locations with spaces in the names.

Previously, if you used a city name like 'New York' it would only set
the third argument as 'New' and 'York' was never considered.

An additional change is the use of command= in place of cmd= for the
`ensure_single_process` script. This is necessary as `ps` on MacOS
doesn't support `-o cmd=` as a filter for the output and instead prefers
`-o command`. This new option also works the same on Linux's `ps` command.